fix:修复了1）fmt::format对armor.name的识别问题，将其直接输出为int.2)tinympc的CMaKeLists的库链接

+++ app/auto_aim/detector.cpp
@@ -351,7 +351,7 @@ cv::Point2f Detector::get_center_norm(const cv::Mat& bgr_img, const cv::Point2f&
 
 void Detector::save(const Armor& armor) const {
     auto file_name = fmt::format("{:%Y-%m-%d_%H-%M-%S}", std::chrono::system_clock::now());
-    auto img_path = fmt::format("{}/{}_{}.jpg", save_path_, armor.name, file_name);
+    auto img_path = fmt::format("{}/{}_{}.jpg", save_path_, static_cast<int>(armor.name), file_name);
     cv::imwrite(img_path, armor.pattern);
 }
